+++ internal/test/test_backend.go
@@ -32,6 +32,9 @@ func NewBackend() *Backend {
 	b.mux.HandleFunc("/ws", echo)
 	b.mux.HandleFunc("/pdf", pdf)
 	b.mux.HandleFunc("/small", small)
+	b.mux.HandleFunc("/error", func(w http.ResponseWriter, _ *http.Request) {
+		w.WriteHeader(http.StatusInternalServerError)
+	})
 
 	return b
 }