+++ CHANGELOG.md
@@ -1,10 +1,23 @@
 # Couper Changelog
 
-## [Unreleased](https://github.com/avenga/couper/compare/1.3...master)
+## [Unreleased](https://github.com/avenga/couper/compare/1.3.1...master)
 
 Unreleased changes are available as `avenga/couper:edge` container.
 
+* **Added**
+  * OAuth2 Authorization Code Grant Flow: `beta_oauth2 {}` block; `beta_oauth_authorization_url()`, `beta_oauth_code_verifier()` and `beta_oauth_csrf_token()` functions ([#247](https://github.com/avenga/couper/pull/247))
+  * Run option `-accept-forwarded-url` and setting `accept_forwarded_url` to accept `proto`, `host`, or `port` from `X-Forwarded-Proto`, `X-Forwarded-Host` or `X-Forwarded-Port` request headers, affecting new client request variables `request.url`, `request.origin`, `request.protocol`, `request.host` and `request.port` ([#255](https://github.com/avenga/couper/pull/255))
+
 ---
+
+## [1.3.1](https://github.com/avenga/couper/compare/1.3...1.3.1)
+
+* **Changed**
+  * `Error` log-level for upstream responses with status `500` to `Info` log-level ([#258](https://github.com/avenga/couper/pull/258))
+
+* **Fixed**
+  * Missing support for `set_response_status` within a plain `error_handler` block ([#257](https://github.com/avenga/couper/pull/257))
+  * Panic in jwt_sign() and saml_sso_url() functions without proper configuration ([#243](https://github.com/avenga/couper/issues/243))
 
 ## [1.3](https://github.com/avenga/couper/compare/1.2...1.3)
 
@@ -14,15 +27,13 @@ Unreleased changes are available as `avenga/couper:edge` container.
 
 * **Changed**
   * Stronger configuration check for `path` and `path_prefix` attributes, possibly resulting in configuration errors ([#232](https://github.com/avenga/couper/pull/232))
-  * Modifier (`set/add/remove_response_headers`) is available for `api`, `files`, `server` and `spa` block too ([#248](https://github.com/avenga/couper/pull/248))
+  * Modifier (`set/add/remove_response_headers`) is available for `api`, `files`, `server` and `spa` block, too ([#248](https://github.com/avenga/couper/pull/248))
 
 * **Fixed**
   * The `path` field in the backend log ([#232](https://github.com/avenga/couper/pull/232))
   * Upstream requests with a known body-size have a `Content-Length` HTTP header field instead of `Transfer-Encoding: chunked` ([#163](https://github.com/avenga/couper/issues/163))
   * Exit endpoint if an error is occurred in `request` or `proxy` instead of processing a defined `response` ([#233](https://github.com/avenga/couper/pull/233))
   * No GZIP compression for small response bodies ([#186](https://github.com/avenga/couper/issues/186))
-
----
 
 ## [1.2](https://github.com/avenga/couper/compare/1.1.1...1.2)